requirements.txt
overrides
absl-py==0.9.0
alabaster==0.7.12
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
docutils==0.16
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
imagesize==1.2.0
importlib-metadata==1.6.0
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
latexcodec==2.0.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==8.2.0
moviepy==1.0.2
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
oset==0.1.3
packaging==20.1
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
proglog==0.1.9
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybtex==0.22.2
pybtex-docutils==0.2.2
Pygments==2.5.2
pyparsing==2.4.6
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==1.0.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
wrapt==1.12.0
zipp==3.1.0
